layers: Move Video CoreCheck to own file
