+++ BUILD.gn
@@ -187,6 +187,7 @@ core_validation_sources = [
   "layers/synchronization_validation.h",
   "layers/video_session_state.cpp",
   "layers/video_session_state.h",
+  "layers/video_validation.cpp",
   "layers/wsi_validation.cpp",
 ]
