+++ build-android/jni/Android.mk
@@ -50,6 +50,7 @@ LOCAL_SRC_FILES += $(SRC_DIR)/layers/pipeline_validation.cpp
 LOCAL_SRC_FILES += $(SRC_DIR)/layers/queue_state.cpp
 LOCAL_SRC_FILES += $(SRC_DIR)/layers/render_pass_state.cpp
 LOCAL_SRC_FILES += $(SRC_DIR)/layers/video_session_state.cpp
+LOCAL_SRC_FILES += $(SRC_DIR)/layers/video_validation.cpp
 LOCAL_SRC_FILES += $(SRC_DIR)/layers/core_validation.cpp
 LOCAL_SRC_FILES += $(SRC_DIR)/layers/drawdispatch.cpp
 LOCAL_SRC_FILES += $(SRC_DIR)/layers/descriptor_sets.cpp
